requirements.txt
sqlalchemy>=2.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0 
python-dotenv>=1.0.0
jinja2>=3.1.0
pandas>=1.5.0
aiofiles>=23.1.0
python-multipart>=0.0.5
requests>=2.28.0